+++ source/Tests/Infrastructure/OutgoingMessages/AggregationResult/SampleData.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Tests.Infrastructure.OutgoingMessages.NotifyAggreagtedMeasureData;
+namespace Tests.Infrastructure.OutgoingMessages.AggregationResult;
 
 internal static class SampleData
 {
